File Split
This branch ended up being a file split for easier readability along with some additional functions for the initialization class.
