+++ enemyai.py
@@ -1,9 +1,8 @@
-#Files for functions, classes, and methods used by the enemy for
-#decision making.
+#Files for functions, classes, and methods used to allow for expanded
+#enemy mech features and provide decision making to the enemy mech.
 #################################80CHAR#########################################
 import random
 import dreadnought
-import menu
 import playerGrid
 import weapons
 
@@ -27,14 +26,13 @@ class BadMech(Dreadnought):
                 	"repairing",
                 ]
                 mechLevel = 1
-	def Level_Up(self, level):
-		if level in range(1,3):
-			self.HealthPoints = self.HealthPoints * level
-			self.mechLevel = level
-		else:
-			level = (self.mechLevel)+1
-			self.HealthPoints = self.HealthPoints * level
-			self.mechLevel = level
+	def Set_Level(self, level):
+		self.HealthPoints = self.HealthPoints * level
+		self.mechLevel = level
+	def Level_Up(self):
+		level = self.mechLevel+1
+		self.HealthPoints = self.HealthPoints * level
+		self.mechLevel = level
 	def Weapon_Change(self, hand):
 		if hand:
 			self.leftWeapon = new Random_Weapon()
@@ -58,49 +56,6 @@ class BadMech(Dreadnought):
 		else:
 			self.mechStatus = "normal"
 
-
-def Random_Weapon():
-	choice = random.randint(1, 4)
-	if choice == 1:
-		return new Bolter()
-	elif choice == 2:
-		return new MeleeWeapon1()
-	elif choice == 3:
-		return new AOEWeapon1()
-	else:
-		return new AssaultCannon()
-
-class StateModel:
-	playerMech
-	playerPosition
-	enemyMech
-	enemyPosition
-	percent_to_move
-	percent_to_defend
-	def __init__(self):
-		#construct player and 9 enemies
-		self.playerMech = new Dreadnought(
-			"template", 1, new Bolter(), new Bolter()
-			)
-		self.playerPosition = new playerGrid()
-		self.enemyMech = []
-		self.enemyPosition = []
-		self.percent_to_move = []
-		self.percent_to_defend = []
-		for i in range(9):
-			self.enemyMech.append(
-				new BadMech(10, Random_Weapon(), Random_Weapon())
-				)
-			if 0 <= i <= 2:
-				self.enemyMech[i].Level_Up(1)
-			if 3 <= i <= 5:
-				self.enemyMech[i].Level_Up(2)
-			if 6 <= i <= 8:
-				self.enemyMech[i].Level_Up(3)
-			self.enemyPosition.append(new playerGrid())
-			self.percent_to_move.append(33)
-			self.percent_to_defend.append(10)
-
 class DecisionMaker:
 	#declare variables that should be initialized
 	world_state
@@ -111,8 +66,8 @@ class DecisionMaker:
 		self.world_state = world
 		self.game_difficulty = difficulty
 	def Take_Action(self):
-		defend = self.world_state.percent_to_defend[current_enemy]
-		move = self.world_state.percent_to_move[current_enemy]
+		defend = 33
+		move = 10
 		#on random chance will defend
 		if random.randint(1, 100) <= defend:
 			self._Defense_Action()
@@ -146,11 +101,3 @@ class DecisionMaker:
 		#if main is best return true
 		enemy = self.enemyMech
 		return enemy.leftWeapon.damage >= enemy.leftWeapon.damage
-	def Change_Enemy(self, round, choice):
-		#choice is 1-3 but based on the round will pick enemy 0-8
-		if round == 1:
-			self.current_enemy = choice - 1
-		elif round == 2:
-			self.current_enemy = choice + 2
-		else:
-			self.current_enemy = choice + 5
